Use Uint64 for SDL_GetTicks counters in id_in/sd/vl/wl_draw
SDL3's SDL_GetTicks returns Uint64. Storing in Uint32 silently truncates
the high bits; the value wraps after ~49 days of uptime which would
break the ackStartTime timeout check, music-tick rate-limiting, frame
present pacing, and CalcTics elapsed-ms calculation. Practical impact
is small (sessions rarely last that long) but it's a real bug — and
fixing it silences several C4244 Uint64-to-Uint32 warnings.

+++ src/id_in.c
@@ -33,7 +33,7 @@ word       DemoOffset, DemoSize;
 static int   mouseButtons;
 static int   mouseDX, mouseDY;
 static boolean ackStarted;
-static Uint32 ackStartTime;
+static Uint64 ackStartTime;
 static void (*keyHook)(void);
 
 // Forward declarations
